pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-first-runtime"
version = "0.1.0"
description = "AI-First Reference Runtime - Execute capability specs safely with undo support"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "AI-First Team", email = "team@ai-first.dev"}
]
keywords = ["ai", "runtime", "mcp", "capability", "safety", "undo"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "click>=8.1",
    "rich>=13.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.5",
]
mcp = [
    "mcp>=1.0",
]

[project.scripts]
airun = "runtime.cli:main"

[project.urls]
Homepage = "https://github.com/gmood2008/ai-first-runtime"
Documentation = "https://github.com/gmood2008/ai-first-runtime/blob/main/README.md"
Repository = "https://github.com/gmood2008/ai-first-runtime"
Issues = "https://github.com/gmood2008/ai-first-runtime/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src/runtime --cov-report=term-missing"
